refactor(ci): satisfy native ownership quality gates
