+++ src/codex_plugin_scanner/guard/native_pretool.py
@@ -12,7 +12,7 @@ from pathlib import Path
 from typing import Any
 
 from .codex_hook_launch_runtime import run_isolated_hook_process
-from .native_route_receipt import record_native_hook_route
+from .native_route_receipt import record_native_hook_result
 from .native_runtime import _isolated_environment, _native_error, native_runtime_status
 from .native_runtime_resident import resident_native_request
 from .native_runtime_resilience import (
@@ -50,11 +50,9 @@ def _decode_pre_tool(payload: object, *, command: str) -> dict[str, Any] | None:
         or not isinstance(model, dict)
         or model.get("normalized_text") != command.strip()
     ):
-        record_native_hook_route("native_fail_safe")
-        return None
+        return record_native_hook_result("native_fail_safe", None)
     if explicitly_benign != (decision == "allow" and action == "allow"):
-        record_native_hook_route("native_fail_safe")
-        return None
+        return record_native_hook_result("native_fail_safe", None)
     return payload
 
 
@@ -79,7 +77,7 @@ def review_pre_tool_native(
         or timeout_seconds <= 0
     ):
         if status.mode in {"auto", "force"}:
-            record_native_hook_route("native_fail_safe")
+            return record_native_hook_result("native_fail_safe", None)
         return None
     request = {
         "command": command,
@@ -89,8 +87,7 @@ def review_pre_tool_native(
     }
     encoded = json.dumps(request, separators=(",", ":"), ensure_ascii=False).encode("utf-8")
     if len(encoded) > _MAX_REQUEST_BYTES:
-        record_native_hook_route("native_fail_safe")
-        return None
+        return record_native_hook_result("native_fail_safe", None)
     timeout_seconds = min(timeout_seconds, 1.0)
     environment = _isolated_environment()
     if _RESIDENT_PROTOCOL_FEATURE in status.capabilities.features:
@@ -114,13 +111,11 @@ def review_pre_tool_native(
                 payload = None
             if _native_error(payload) == "native_overloaded":
                 native_record_overload(status.identity.sha256, guard_home)
-                record_native_hook_route("native_fail_safe")
-                return None
+                return record_native_hook_result("native_fail_safe", None)
             decoded = _decode_pre_tool(payload, command=command)
             if decoded is not None:
                 native_record_resident_success(status.identity.sha256, guard_home)
-                record_native_hook_route("native_resident")
-                return decoded
+                return record_native_hook_result("native_resident", decoded)
         native_record_resident_failure(
             status.identity.sha256,
             guard_home,
@@ -128,8 +123,7 @@ def review_pre_tool_native(
         )
     with native_oneshot_lease(status.identity.sha256, guard_home) as acquired:
         if not acquired:
-            record_native_hook_route("native_fail_safe")
-            return None
+            return record_native_hook_result("native_fail_safe", None)
         result = run_isolated_hook_process(
             (str(status.identity.path), "pre-tool", "--stdin"),
             input_text=encoded.decode("utf-8"),
@@ -144,8 +138,7 @@ def review_pre_tool_native(
                 guard_home,
                 reason="native_pre_tool_oneshot_failed",
             )
-            record_native_hook_route("native_fail_safe")
-            return None
+            return record_native_hook_result("native_fail_safe", None)
         try:
             decoded = _decode_pre_tool(json.loads(result.stdout), command=command)
         except json.JSONDecodeError:
@@ -156,11 +149,9 @@ def review_pre_tool_native(
                 guard_home,
                 reason="native_pre_tool_oneshot_invalid",
             )
-            record_native_hook_route("native_fail_safe")
-            return None
+            return record_native_hook_result("native_fail_safe", None)
         native_record_oneshot_success(status.identity.sha256, guard_home)
-        record_native_hook_route("native_oneshot")
-        return decoded
+        return record_native_hook_result("native_oneshot", decoded)
 
 
 def native_pre_tool_policy_floor(
