+++ src/codex_plugin_scanner/guard/native_route_receipt.py
@@ -3,7 +3,7 @@
 from __future__ import annotations
 
 from contextvars import ContextVar
-from typing import Literal, TypeAlias
+from typing import Literal, TypeAlias, TypeVar
 
 NativeHookRoute: TypeAlias = Literal["native_resident", "native_oneshot", "native_fail_safe"]
 HookDecisionRoute: TypeAlias = Literal[
@@ -17,6 +17,7 @@ _NATIVE_HOOK_ROUTE: ContextVar[HookDecisionRoute | None] = ContextVar(
     "hol_guard_native_hook_route",
     default=None,
 )
+_T = TypeVar("_T")
 
 
 def reset_native_hook_route() -> None:
@@ -29,6 +30,12 @@ def record_native_hook_route(route: NativeHookRoute) -> None:
     _NATIVE_HOOK_ROUTE.set(route)
 
 
+def record_native_hook_result(route: NativeHookRoute, result: _T) -> _T:
+    """Attach route provenance while preserving the decision result's type."""
+    record_native_hook_route(route)
+    return result
+
+
 def record_python_semantic_hook_route() -> None:
     _NATIVE_HOOK_ROUTE.set("python_semantic")
 
@@ -39,6 +46,7 @@ def native_hook_route() -> HookDecisionRoute | None:
 
 __all__ = [
     "native_hook_route",
+    "record_native_hook_result",
     "record_native_hook_route",
     "record_python_semantic_hook_route",
     "reset_native_hook_route",
